+++ packages/embedded-client/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## [Unreleased]
 
-No notable changes.
+### Added
+
+- [`Options.defaultScope`](../../docs/embedded-client.options.defaultscope.md) to configure the embedded policy decision point server's [default scope](https://docs.cerbos.dev/cerbos/latest/configuration/engine#default_scope) ([#1333](https://github.com/cerbos/cerbos-sdk-javascript/pull/1333))
 
 ## [0.3.0] - 2026-01-07